Add files via upload
Upload of Main Code 

This is a spoof SWIFT Terminal which can be used to make a number of transactions.
Save as a .BAT file and run.

To amend window size, open the run the BAT file and right click on the box banner. Select PROPERTIES and enter the box size. 
Next time the BAT file runs it will be at this requested size.

- Fake SWIFT Terminal to make a payment (needs source acc no and sort code, destination acc no and sort code, amount and currency.
- Can also close and cancel payments
- Option to review payment logs gives a random error regardless of what is entered
- Log deletion does not request password etc